feat(model): fp16 mixed precision with fp32 master weights (#27)
Make `precision=fp16` actually drive compute. Until now the MLX model ran
entirely in fp32 regardless of `config.precision`, so a fp16 scale run got
none of the measured speedup and the dynamic loss scaler was a no-op.

Cast params to the compute dtype AT THE MATMUL SITE (fp32 masters preserved):
- heavy GEMMs (in/x/out projections, depthwise conv) run in fp16/bf16;
- the activation/residual stream between layers is the compute dtype;
- RMSNorm reductions, the SSM scan internals (dt_proj/softplus/decays/cumsum/
  _segsum/state einsums), and the wide-vocab LM-head logits + cross-entropy
  stay fp32;
- SSM recurrent state stays fp32.

MLX autodiff returns fp32 grads through the `astype` cast, so AdamW keeps
updating the fp32 masters with no optimizer change and the DynamicLossScaler
(#3) finally does meaningful work. Every cast is guarded so `precision=fp32`
is the original op verbatim — bit-identical (smoke gate + conformance intact).

- bench_precision.py: add `--mode model` end-to-end fwd+bwd benchmark
  (fp16 ~1.08x fp32 on the real poc model); default `gemm` mode unchanged.
- tests: new tests/test_mlx_precision.py (dtype policy, fp32 bit-identity,
  fp16 short-run convergence with scaler, grad-checkpoint determinism).

Verified: full pytest (43) green, smoke gate exact (fp32 resume), fp32
conformance ~1e-4 intact.

+++ src/model/mlx_train_step.py
@@ -40,7 +40,11 @@ def make_train_step(model, optimizer, *, grad_clip: float = 1.0,
         logits = model.forward(inputs)                      # (B, L, V)
         V = logits.shape[-1]
         t = mx.array(targets).reshape(-1).astype(mx.int32)
-        ce = nn.losses.cross_entropy(logits.reshape(-1, V), t, reduction="mean")
+        # Cross-entropy in fp32 (wide-vocab softmax stability). The MLX backend's
+        # `_head` already returns fp32 logits, so this is a no-op there; the cast
+        # keeps the contract explicit and backend-independent.
+        ce = nn.losses.cross_entropy(logits.reshape(-1, V).astype(mx.float32),
+                                     t, reduction="mean")
         return ce * scaler.scale if scaler else ce
 
     loss_and_grad = nn.value_and_grad(model, loss_fn)
